oo_inherit_simple2 test added
This test is equivalent to oo_inherit.py, except that *args (that are not
supported yet) it uses default arguments. This test should mainly test the
inheritance, so *args are not important here. This oo_inherit_simple2 still
fails, until we can call the __init__ method from the parent class.

The original test was left intact. That will only start working after we fix
the oo_inherit_simple2 and also implement *args.